+++ A02YYUW.h
@@ -16,21 +16,17 @@ constexpr int A02YY_OK          = 0;
 constexpr int A02YY_ERR_CRC     = -100;
 constexpr int A02YY_ERR_TIMEOUT = -101;
 
-class A02YYUW
+
+////////////////////////////////////////////
+//
+//  BASE CLASS
+//
+class A02YY
 {
 public:
-  A02YYUW(Stream * str, uint8_t TX);
+  A02YY(Stream * str);
 
   void     begin();
-
-  //  needed for the UART AUTO only
-  //  true  = processed mode
-  //  false = real time mode (default)
-  void     setProcessingMode(bool mode);
-  bool     getProcessingMode();
-
-  //  needed for the UART CONTROLLED only
-  void     request();
 
   //  needed by all
   bool     newDistance();
@@ -58,30 +54,54 @@ protected:
 
 ////////////////////////////////////////////
 //
-//  DERIVED CLASSES (experimental wrappers)
+//  DERIVED CLASSES - UART AUTO - A02YYU(W)
 //
-class A02YYU : public A02YYUW
+class A02YYU : public A02YY
 {
   public:
-  A02YYU(Stream * str, uint8_t TX) : A02YYUW(str, TX)
+  A02YYU(Stream * str, uint8_t TX) : A02YY(str)
   {
-  }
+    _TX = TX;
+    //  needed? as Serial implies TX == OUTPUT.
+    pinMode(_TX, OUTPUT);
+  };
+
+  //  true  = processed mode
+  //  false = real time mode (default)
+  void setProcessingMode(bool mode);
+  bool getProcessingMode();
 };
 
-class A02YYT : public A02YYUW
+//  wrapper only
+class A02YYUW : public A02YYU
 {
   public:
-  A02YYT(Stream * str, uint8_t TX) : A02YYUW(str, TX)
+  A02YYUW(Stream * str, uint8_t TX) : A02YYU(str, TX)
   {
-  }
+  };
 };
 
-class A02YYTW : public A02YYUW
+
+////////////////////////////////////////////
+//
+//  DERIVED CLASSES - UART CONTROLLED - A02YYT(W)
+//
+class A02YYT : public A02YY
 {
   public:
-  A02YYTW(Stream * str, uint8_t TX) : A02YYUW(str, TX)
+  A02YYT(Stream * str) : A02YY(str)
   {
-  }
+  };
+  void request();
+};
+
+//  wrapper only
+class A02YYTW : public A02YYT
+{
+  public:
+  A02YYTW(Stream * str) : A02YYT(str)
+  {
+  };
 };